WIP: moved  example_driver from internal -> testing_util
