+++ google/cloud/bigtable/examples/CMakeLists.txt
@@ -84,6 +84,7 @@ foreach (fname ${bigtable_examples})
                 bigtable_client
                 bigtable_protos
                 google_cloud_cpp_common
+                google_cloud_cpp_testing
                 google_cloud_cpp_grpc_utils
                 gRPC::grpc++
                 gRPC::grpc
