+++ google/cloud/google_cloud_cpp_common.bzl
@@ -32,7 +32,6 @@ google_cloud_cpp_common_hdrs = [
     "internal/diagnostics_push.inc",
     "internal/disable_msvc_crt_secure_warnings.inc",
     "internal/disjunction.h",
-    "internal/example_driver.h",
     "internal/filesystem.h",
     "internal/format_time_point.h",
     "internal/future_base.h",
@@ -67,7 +66,6 @@ google_cloud_cpp_common_srcs = [
     "iam_policy.cc",
     "internal/backoff_policy.cc",
     "internal/compiler_info.cc",
-    "internal/example_driver.cc",
     "internal/filesystem.cc",
     "internal/format_time_point.cc",
     "internal/future_impl.cc",
