+++ google/cloud/pubsub/samples/BUILD
@@ -27,6 +27,7 @@ cc_library(
     deps = [
         "//google/cloud:google_cloud_cpp_common",
         "//google/cloud/pubsub:pubsub_client",
+        "//google/cloud/testing_util:google_cloud_cpp_testing",
     ],
 )
