+++ google/cloud/pubsub/samples/pubsub_samples_common.h
@@ -17,7 +17,7 @@
 
 #include "google/cloud/pubsub/publisher_client.h"
 #include "google/cloud/pubsub/subscriber_client.h"
-#include "google/cloud/internal/example_driver.h"
+#include "google/cloud/testing_util/example_driver.h"
 
 namespace google {
 namespace cloud {
