+++ google/cloud/storage/examples/CMakeLists.txt
@@ -86,8 +86,9 @@ endif ()
 foreach (fname ${storage_examples})
     google_cloud_cpp_add_executable(target "storage_examples" "${fname}")
     target_link_libraries(
-        ${target} PRIVATE storage_examples_common storage_client_grpc
-                          storage_client google_cloud_cpp_common)
+        ${target}
+        PRIVATE storage_examples_common storage_client_grpc storage_client
+                google_cloud_cpp_common google_cloud_cpp_testing)
     add_test(NAME ${target} COMMAND ${target})
     set_tests_properties(${target} PROPERTIES LABELS "integration-tests")
     set_tests_properties(
