+++ google/cloud/testing_util/CMakeLists.txt
@@ -25,6 +25,8 @@ if (BUILD_TESTING OR GOOGLE_CLOUD_CPP_TESTING_UTIL_ENABLE_INSTALL)
         capture_log_lines_backend.h
         check_predicate_becomes_false.h
         chrono_literals.h
+        example_driver.cc
+        example_driver.h
         expect_exception.h
         expect_future_error.h
         scoped_environment.cc
@@ -39,7 +41,7 @@ if (BUILD_TESTING OR GOOGLE_CLOUD_CPP_TESTING_UTIL_ENABLE_INSTALL)
 
     set(google_cloud_cpp_testing_unit_tests
         # cmake-format: sort
-        assert_ok_test.cc scoped_environment_test.cc)
+        assert_ok_test.cc example_driver_test.cc scoped_environment_test.cc)
 
     # Export the list of unit tests so the Bazel BUILD file can pick it up.
     export_list_to_bazel("google_cloud_cpp_testing_unit_tests.bzl"
