+++ google/cloud/testing_util/example_driver.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/internal/example_driver.h"
+#include "google/cloud/testing_util/example_driver.h"
 #include "google/cloud/internal/getenv.h"
 #include <iostream>
