+++ google/cloud/testing_util/google_cloud_cpp_testing_unit_tests.bzl
@@ -18,5 +18,6 @@
 
 google_cloud_cpp_testing_unit_tests = [
     "assert_ok_test.cc",
+    "example_driver_test.cc",
     "scoped_environment_test.cc",
 ]
